[0.6.7] feat(**) Added filter exact & similitude parameters to Find, changed skip & take from int to long, genre tree with addable, next & previous for search results

+++ Musicx.Application.Web/Interfaces/UseCases/Specifics/IGetAlbumAttributesByAlbum.cs
@@ -6,6 +6,6 @@ namespace Musicx.Application.Web.Interfaces.UseCases.Specifics;
 public interface IGetAlbumAttributesByAlbum
 {
     Task<OutGenericList<OutUserAlbumAttribute>> ExecuteAsync(long albumId, 
-        int skip = 0, int take = 100, CancellationToken cancellationToken = default);
-    OutGenericList<OutUserAlbumAttribute> Execute(long albumId, int skip = 0, int take = 100);
+        long skip = 0, long take = 100, CancellationToken cancellationToken = default);
+    OutGenericList<OutUserAlbumAttribute> Execute(long albumId, long skip = 0, long take = 100);
 }
