+++ Musicx.Application.Web/Interfaces/UseCases/Specifics/IGetAlbumAttributesByUser.cs
@@ -6,6 +6,6 @@ namespace Musicx.Application.Web.Interfaces.UseCases.Specifics;
 public interface IGetAlbumAttributesByUser
 {
     Task<OutGenericList<OutUserAlbumAttribute>> ExecuteAsync(long userId, 
-        int skip = 0, int take = 100, CancellationToken token = default);
-    OutGenericList<OutUserAlbumAttribute> Execute(long userId, int skip = 0, int take = 100);
+        long skip = 0, long take = 100, CancellationToken token = default);
+    OutGenericList<OutUserAlbumAttribute> Execute(long userId, long skip = 0, long take = 100);
 }
